Publish event when adjusting the dataflow engine replicas
